// Qanat.Web/src/app/shared/generated/model/open-et-sync-dto.ts
export class OpenETSyncDto { 
    OpenETSyncID?: number;
    Geography?: GeographyDto;
    OpenETDataType?: OpenETDataTypeSimpleDto;
    Year?: number;
    Month?: number;
    LastSyncDate?: string;
    LastSyncStatus?: OpenETSyncResultTypeSimpleDto;
    LastSuccessfulSyncDate?: string;
    LastSyncMessage?: string;
    HasInProgressSync?: boolean;
    LastRasterCalculationDate?: string;
    LastRasterCalculationStatus?: OpenETRasterCalculationResultTypeSimpleDto;
    LastSuccessfulCalculationDate?: string;
    LastRasterCalculationMessage?: string;
    FileResourceGUID?: string;
    FileResourceOriginalName?: string;
    FileResourceOriginalFileExtension?: string;
    FinalizeDate?: string;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}